requirements.txt
flask>=2.3.3
pillow>=10.0.0
python-dotenv==1.0.0
werkzeug>=2.3.7
blinker>=1.6.2
transformers>=4.30.0
torch>=2.0.0
diffusers>=0.20.0
accelerate>=0.20.0
SpeechRecognition>=3.8.0